Fix more CMake compiler warnings about implicit data conversion
Bug: 436300898
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262734
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>

+++ src/dawn/native/BindGroup.cpp
@@ -125,7 +125,7 @@ MaybeError ValidateBufferBinding(const DeviceBase* device,
             DAWN_UNREACHABLE();
     }
 
-    DAWN_INVALID_IF(!IsAligned(entry.offset, requiredBindingAlignment),
+    DAWN_INVALID_IF(!IsAligned(static_cast<uint32_t>(entry.offset), requiredBindingAlignment),
                     "Offset (%u) of %s does not satisfy the minimum %s alignment (%u).",
                     entry.offset, entry.buffer, layout.type, requiredBindingAlignment);
