+++ src/dawn/native/Commands.cpp
@@ -422,7 +422,7 @@ void SkipCommand(CommandIterator* commands, Command type) {
 
 const char* AddNullTerminatedString(CommandAllocator* allocator, StringView s, uint32_t* length) {
     std::string_view view = s;
-    *length = view.length();
+    *length = static_cast<uint32_t>(view.length());
 
     // Include extra null-terminator character. The string_view may not be null-terminated. It also
     // may already have a null-terminator inside of it, in which case adding the null-terminator is
