+++ src/dawn/native/ComputePipeline.cpp
@@ -41,13 +41,14 @@ MaybeError ValidateComputePipelineDescriptor(DeviceBase* device,
     }
 
     ShaderModuleEntryPoint entryPoint;
-    DAWN_TRY_ASSIGN_CONTEXT(entryPoint,
-                            ValidateProgrammableStage(
-                                device, descriptor->compute.module, descriptor->compute.entryPoint,
-                                descriptor->compute.constantCount, descriptor->compute.constants,
-                                descriptor->layout, SingleShaderStage::Compute),
-                            "validating compute stage (%s, entryPoint: %s).",
-                            descriptor->compute.module, descriptor->compute.entryPoint);
+    DAWN_TRY_ASSIGN_CONTEXT(
+        entryPoint,
+        ValidateProgrammableStage(
+            device, descriptor->compute.module, descriptor->compute.entryPoint,
+            static_cast<uint32_t>(descriptor->compute.constantCount), descriptor->compute.constants,
+            descriptor->layout, SingleShaderStage::Compute),
+        "validating compute stage (%s, entryPoint: %s).", descriptor->compute.module,
+        descriptor->compute.entryPoint);
     return {};
 }
